test(oprish): move tests to their appropriate files
