+++ oprish/src/routes/mod.rs
@@ -26,3 +26,35 @@ pub async fn get_instance_info(
 pub fn get_routes() -> Vec<Route> {
     routes![get_instance_info]
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::{rocket};
+    use rocket::{http::Status, local::asynchronous::Client};
+    use todel::{
+        models::{InstanceInfo},
+        Conf,
+    };
+
+    #[rocket::async_test]
+    async fn index() {
+        let client = Client::untracked(rocket().unwrap()).await.unwrap();
+        let conf = &client.rocket().state::<Conf>().unwrap();
+
+        let response = client.get("/").dispatch().await;
+
+        assert_eq!(response.status(), Status::Ok);
+        assert_eq!(
+            response.into_string().await.unwrap(),
+            serde_json::to_string(&InstanceInfo::from_conf(conf, false)).unwrap()
+        );
+
+        let response = client.get("/?rate_limits").dispatch().await;
+
+        assert_eq!(response.status(), Status::Ok);
+        assert_eq!(
+            response.into_string().await.unwrap(),
+            serde_json::to_string(&InstanceInfo::from_conf(conf, true)).unwrap()
+        );
+    }
+}
